move variables into generatePassword function
moved some the variables that were global in the class-time code and
mmoved them into the generatePassword function. This solves some issues
including the string for usable characters not reseting, and I changed
the passwrod variable to passWord in the function to avoid future
conflict with the password variable in writePassword function. Also,
moving the variables inside the funciton means that the prompts only
show when the button is pressed, which is both dope and the intended
behavior.