add input scrubber for passwordLength
adds a function to gather the desired password length, and checks that
it is between 8 and 128, inclusive. If the user passes a string, the
string's length is taken as the input and then checked that it is
between 8 and 128, inclusive. If the user hits cancel the first time, it
will prompt once more, and cancelling again will exit the password
generation task. If the user puts in a bad input first, and then cancels
the second time they are prompted, it will again cancel the task.